% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bovids.R
\docType{data}
\name{bovids}
\alias{bovids}
\title{13 orthologous mtDNA-encoded proteins of 11 bovine species.}
\format{
A data frame with 13 rows (one per protein) and 11 variables (one per species)
}
\source{
\url{https://www.ncbi.nlm.nih.gov/genome/organelle/}
}
\usage{
bovids
}
\description{
A dataset containing the sequences of mtDNA-encoded proteins from different bovids
}
\keyword{datasets}
