% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{d.phy2df}
\alias{d.phy2df}
\title{Convert a Phylip Distance Matrix into a DataFrame}
\usage{
d.phy2df(phyfile, as = 'matrix')
}
\arguments{
\item{phyfile}{path to the file containing the distances in phylip format.}

\item{as}{class of the output R data. It should be either 'dataframe' or 'matrix'.}
}
\value{
Either a dataframe or a matrix containing the distances read from the phy file.
}
\description{
Converts a phylip distance matrix into a either an R dataFrame or matrix.
}
\examples{
\dontrun{d.phy2df(phyfile = "./data_t/d_dummy.txt")}
}
\seealso{
d.df2pny()
}
