% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdnj.R
\name{ncdnj}
\alias{ncdnj}
\title{Compute a Distance Matrix Using Normalized Compression Distance}
\usage{
ncdnj(wd)
}
\arguments{
\item{wd}{character string indicating the path to the directory where the input files can be found (see details).}
}
\value{
A list where the first element is a symmetric distance matrix and the second one is a phylogenetic tree build using NJ.
}
\description{
Computes a distance matrix using normalized compression distance.
}
\details{
The input files, which must be found at the wd provided, consist of a file named 'list.txt' containing the names of the fasta files to be analyzed (one per line). The referred fasta files also must be found at the provided wd. This function use zpaq to compress files. Thus, the zpaq software must be installed on your system and in the search path for executables if you wish to use this function.
}
\examples{
\donttest{try(ncdnj("./data_t"))}
}
\seealso{
ncd()
}
