% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reyes.R
\docType{data}
\name{reyes}
\alias{reyes}
\title{13 orthologous mtDNA-encoded proteins of 34 mammalian species.}
\format{
A data frame with 13 rows (one per protein) and 34 variables (one per species)
}
\source{
\url{https://www.ncbi.nlm.nih.gov/genome/organelle/}
}
\usage{
reyes
}
\description{
A dataset containing the sequences of mtDNA-encoded proteins from different mammalian species (Reyes et al. 1996, Mol.Biol. Evol. 17:979)
}
\keyword{datasets}
